// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  FARMER: 'FARMER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AuthProvider = {
  EMAIL: 'EMAIL',
  GOOGLE: 'GOOGLE'
} as const

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider]


export const AreaUnit = {
  ACRE: 'ACRE',
  HECTARE: 'HECTARE'
} as const

export type AreaUnit = (typeof AreaUnit)[keyof typeof AreaUnit]


export const Region = {
  CENTRAL: 'CENTRAL',
  NORTH: 'NORTH',
  EAST: 'EAST',
  WEST: 'WEST',
  SOUTH: 'SOUTH'
} as const

export type Region = (typeof Region)[keyof typeof Region]


export const SoilType = {
  SANDY: 'SANDY',
  CLAY: 'CLAY',
  LOAMY: 'LOAMY',
  SILTY: 'SILTY',
  PEATY: 'PEATY',
  CHALKY: 'CHALKY',
  OTHER: 'OTHER'
} as const

export type SoilType = (typeof SoilType)[keyof typeof SoilType]


export const IrrigationType = {
  RAINFED: 'RAINFED',
  CANAL: 'CANAL',
  TUBE_WELL: 'TUBE_WELL',
  WELL: 'WELL',
  DRIP: 'DRIP',
  SPRINKLER: 'SPRINKLER',
  OTHER: 'OTHER'
} as const

export type IrrigationType = (typeof IrrigationType)[keyof typeof IrrigationType]


export const Season = {
  RABI: 'RABI',
  KHARIF: 'KHARIF',
  ZAID: 'ZAID'
} as const

export type Season = (typeof Season)[keyof typeof Season]


export const CropStatus = {
  PLANNED: 'PLANNED',
  ACTIVE: 'ACTIVE',
  HARVESTED: 'HARVESTED',
  FAILED: 'FAILED'
} as const

export type CropStatus = (typeof CropStatus)[keyof typeof CropStatus]


export const CropType = {
  WHEAT: 'WHEAT',
  RICE: 'RICE',
  MAIZE: 'MAIZE',
  COTTON: 'COTTON',
  SUGARCANE: 'SUGARCANE',
  POTATO: 'POTATO'
} as const

export type CropType = (typeof CropType)[keyof typeof CropType]


export const NotificationType = {
  WEATHER: 'WEATHER',
  RAIN: 'RAIN',
  TEMPERATURE: 'TEMPERATURE',
  WIND: 'WIND',
  HUMIDITY: 'HUMIDITY',
  CROP: 'CROP',
  SOIL_MOISTURE: 'SOIL_MOISTURE',
  SYSTEM: 'SYSTEM'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const NotificationPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH'
} as const

export type NotificationPriority = (typeof NotificationPriority)[keyof typeof NotificationPriority]


export const DevicePlatform = {
  ANDROID: 'ANDROID',
  IOS: 'IOS',
  WEB: 'WEB'
} as const

export type DevicePlatform = (typeof DevicePlatform)[keyof typeof DevicePlatform]